Add files via upload
The user gets to consult a specialist according to his choice for the categories, mentioned on the landing page. The admin of this website can modify, add or remove the list of consultants based upon the requirement. There is a page for consultants where they can add their patient data(which gets stored in the database). This landing page is not complete and neither the contact page. This is just a structure of the interface explained, having full functionality.